% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_resampling.R
\name{bootstrap_resampling}
\alias{bootstrap_resampling}
\title{Bootsrap Resampling}
\usage{
bootstrap_resampling(data, target, times = 10, algorithm = "knn",
  metric = "Accuracy", ...)
}
\arguments{
\item{data}{Dataframe that will be used}

\item{target}{Integer with the column number of the target variable}

\item{times}{Number of times to create resamples}

\item{algorithm}{Only "knn" available}

\item{metric}{Metric that will be evaluated. Possible values: "Accuracy", "Sensitivity" or "Specificity"}
}
\value{
Returns the mean values for the metric selected
}
\description{
Bootsrap Resampling for knn Classifier algorithm
}
\examples{
data<-boostrap_resampling(data = iris, target = 5, times = 10, metric = "Sensitivity")
}
\author{
Alberto Almui?a
}
