% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds.R
\name{create_folds}
\alias{create_folds}
\title{Create Folds}
\usage{
create_folds(data, k = 10)
}
\arguments{
\item{data}{Dataframe that will be used}

\item{k}{Number of sequences created with the rows of the dataset}
}
\value{
Returns a list with 'k' folds
}
\description{
This function creates 'k' folds with length(df)/k rows
}
\examples{
data<-create_folds(iris, 10)
}
\author{
Alberto Almui?a
}
