% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multi_folds.R
\name{create_multi_folds}
\alias{create_multi_folds}
\title{Create Multi Folds}
\usage{
create_multi_folds(data, k = 10, times = 5)
}
\arguments{
\item{data}{Dataframe that will be used}

\item{k}{Number of sequences created with length(df)/k rows}

\item{times}{Number of repeats of "create_folds"}
}
\value{
Returns a list with 'k' sequences 'times' times.
}
\description{
This function creates a resampled dataframe
}
\examples{
data<-create_multi_folds(iris, 10, 5)
}
\author{
Alberto Almui?a
}
