% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_fold_cross_validation.R
\name{k_fold_cross_validation}
\alias{k_fold_cross_validation}
\title{k-Fold Cross Validation}
\usage{
k_fold_cross_validation(data, target, method = c("k_fold",
  "multi_k_fold"), algorithm = "knn", k_fold = 10, times = 5,
  metric = "Accuracy", ...)
}
\arguments{
\item{data}{Dataframe that will be used}

\item{target}{Integer with the column number of the target variable}

\item{method}{Could be "k_fold" or "multi_k_fold"}

\item{algorithm}{Only "knn" available}

\item{k_fold}{Number of folds that will be created}

\item{times}{Number of times to create "multi_k_fold"}

\item{metric}{Metric that will be evaluated. Possible values: "Accuracy", "Sensitivity" or "Specificity"}
}
\value{
Returns the mean values for the metric selected
}
\description{
k-Fold Cross Validation for knn Classifier algorithm
}
\examples{
data<-k_fold_cross_validation(data = iris, target = 5, method = "k_fold", k_fold = 3, metric = "Sensitivity")
}
\author{
Alberto Almui?a
}
