% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_knn_tester.R
\name{k_knn_tester}
\alias{k_knn_tester}
\title{k_knn_tester}
\usage{
k_knn_tester(data, target = is.integer(), k = is.vector(),
  method = c("bootstrap", "cross_validation"), metric = c("Accuracy",
  "Sensitivity", "Specificity"), ...)
}
\arguments{
\item{data}{Dataframe that will be used}

\item{target}{Integer with the column number of the target variable}

\item{k}{vector of values to use}

\item{method}{"bootstrap" or "cross_validation"}

\item{metric}{Metric that will be evaluated. Possible values: "Accuracy", "Sensitivity" or "Specificity"}
}
\value{
Returns the mean values for the metric selected for each 'k' value and plots
}
\description{
Search the best 'k' parameter
}
\examples{
data<-k_knn_tester(data = iris, target = 5, k = c(3,5,7,9), method = "bootstrap", metric = "Sensitivity")
}
\author{
Alberto Almui?a
}
