% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_classifier.R
\name{knn_classifier}
\alias{knn_classifier}
\title{knn Classifier}
\usage{
knn_classifier(data = NULL, p_train = NULL, train = NULL,
  test = NULL, k = 3, target = is.numeric(),
  dist_method = "euclidean", probs = F)
}
\arguments{
\item{data}{Dataframe that will be used}

\item{p_train}{Percentage of rows that will be used as train}

\item{train}{Dataset used as train set (Only when data and p_train are NULL)}

\item{test}{Dataset used as test set (Only when data and p_train are NULL)}

\item{k}{Number of neighbors used}

\item{target}{Integer with the column number of the target variable}

\item{dist_method}{Method to calculate the distance between data. Only "Euclidean" available}

\item{probs}{If TRUE, the results are given in probability. Default is FALSE.}
}
\value{
Returns a list with the original dataset + predicted values and a confusion matrix with several data.
}
\description{
Application of knn algorithm for classification
}
\examples{
data<-knn(data = iris, p_train = 0.7, target = 5, k = 3)
}
\author{
Alberto Almui?a
}
