% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_preprocess.R
\name{knn_preprocess}
\alias{knn_preprocess}
\title{knn Preprocess}
\usage{
knn_preprocess(formula, data, methods = is.character(),
  apply_on = c("numeric", "categorical"), role = c("predictor",
  "outcome"))
}
\arguments{
\item{formula}{Formula that will be evaluated in knn algorithm (for example: Species ~ .)}

\item{data}{Dataframe that will be modified}

\item{methods}{Steps functions of the recipes package to be applied. For example( c("step_center", "step_scale", "step_range"))}

\item{apply_on}{"Categorical" or "Numeric". The column type in which step functions will be applied.}

\item{role}{"Predictor" or "Outcome". Outcome is the target variable and "Predictor" represent all the rest in "formula".}
}
\value{
Returns a data frame with all the changes applied.
}
\description{
This function applies step functions of recipes package
}
\examples{
data<-knn_preprocess(Species ~ ., iris, c("step_center", "step_scale", "step_range"), "Numeric", "Predictor")
data<-knn_preprocess(Species ~ ., iris, c("step_dummy"), "Categorical", "Outcome")
}
\author{
Alberto Almui?a
}
